+++ src/JLD2.jl
@@ -548,6 +548,7 @@ include("data/number_types.jl")
 include("data/custom_serialization.jl")
 include("data/writing_datatypes.jl")
 include("data/reconstructing_datatypes.jl")
+include("data/unknown_types.jl")
 
 include("dataio.jl")
 include("loadsave.jl")
