+++ src/data/type_defs.jl
@@ -12,7 +12,7 @@ struct UnknownType{T}
     fieldnames::Vector{String}
     fieldtypes::Vector{Any}
 
-    UnknownType{T}(name) where T = new(name)
+    UnknownType{T}(name) where T = new(name, [])
     UnknownType{T}(name, parameters) where T = new(name, parameters)
     UnknownType{T}(name, fieldnames, fieldtypes) where T = new(name, [], fieldnames, fieldtypes)
     UnknownType{T}(name, parameters, fieldnames, fieldtypes) where T = new(name, parameters, fieldnames, fieldtypes)
